moved filterComponents to user/nav-layout
